Product Details working on mobile
